Style header and select
