+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\User\\Desktop\\bookstore\\src\\index.js":"1","C:\\Users\\User\\Desktop\\bookstore\\src\\reducers\\filter.js":"2","C:\\Users\\User\\Desktop\\bookstore\\src\\components\\App.js":"3","C:\\Users\\User\\Desktop\\bookstore\\src\\reducers\\books.js":"4","C:\\Users\\User\\Desktop\\bookstore\\src\\containers\\BookList.js":"5","C:\\Users\\User\\Desktop\\bookstore\\src\\containers\\BooksForm.js":"6","C:\\Users\\User\\Desktop\\bookstore\\src\\components\\Book.js":"7","C:\\Users\\User\\Desktop\\bookstore\\src\\components\\CategoryFilter.js":"8","C:\\Users\\User\\Desktop\\bookstore\\src\\actions\\index.js":"9","C:\\Users\\User\\Desktop\\bookstore\\src\\components\\Header.js":"10"},{"size":592,"mtime":1611822628440,"results":"11","hashOfConfig":"12"},{"size":325,"mtime":1611753724126,"results":"13","hashOfConfig":"12"},{"size":384,"mtime":1611821992911,"results":"14","hashOfConfig":"12"},{"size":679,"mtime":1611666508861,"results":"15","hashOfConfig":"12"},{"size":1721,"mtime":1611753724126,"results":"16","hashOfConfig":"12"},{"size":1721,"mtime":1611753724126,"results":"17","hashOfConfig":"12"},{"size":474,"mtime":1611648879378,"results":"18","hashOfConfig":"12"},{"size":1161,"mtime":1611753724126,"results":"19","hashOfConfig":"12"},{"size":277,"mtime":1611753724126,"results":"20","hashOfConfig":"12"},{"size":549,"mtime":1611825975228,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"7e4ybh",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\User\\Desktop\\bookstore\\src\\index.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\reducers\\filter.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\components\\App.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\reducers\\books.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\containers\\BookList.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\containers\\BooksForm.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\components\\Book.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\components\\CategoryFilter.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\actions\\index.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\components\\Header.js",[]]
+[{"C:\\Users\\User\\Desktop\\bookstore\\src\\index.js":"1","C:\\Users\\User\\Desktop\\bookstore\\src\\reducers\\filter.js":"2","C:\\Users\\User\\Desktop\\bookstore\\src\\components\\App.js":"3","C:\\Users\\User\\Desktop\\bookstore\\src\\reducers\\books.js":"4","C:\\Users\\User\\Desktop\\bookstore\\src\\containers\\BookList.js":"5","C:\\Users\\User\\Desktop\\bookstore\\src\\containers\\BooksForm.js":"6","C:\\Users\\User\\Desktop\\bookstore\\src\\components\\Book.js":"7","C:\\Users\\User\\Desktop\\bookstore\\src\\components\\CategoryFilter.js":"8","C:\\Users\\User\\Desktop\\bookstore\\src\\actions\\index.js":"9","C:\\Users\\User\\Desktop\\bookstore\\src\\components\\Header.js":"10"},{"size":592,"mtime":1611822628440,"results":"11","hashOfConfig":"12"},{"size":325,"mtime":1611753724126,"results":"13","hashOfConfig":"12"},{"size":384,"mtime":1611821992911,"results":"14","hashOfConfig":"12"},{"size":679,"mtime":1611666508861,"results":"15","hashOfConfig":"12"},{"size":1279,"mtime":1611828852554,"results":"16","hashOfConfig":"12"},{"size":1721,"mtime":1611753724126,"results":"17","hashOfConfig":"12"},{"size":474,"mtime":1611648879378,"results":"18","hashOfConfig":"12"},{"size":1332,"mtime":1611829220593,"results":"19","hashOfConfig":"12"},{"size":277,"mtime":1611753724126,"results":"20","hashOfConfig":"12"},{"size":656,"mtime":1611829007721,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"7e4ybh",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\User\\Desktop\\bookstore\\src\\index.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\reducers\\filter.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\components\\App.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\reducers\\books.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\containers\\BookList.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\containers\\BooksForm.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\components\\Book.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\components\\CategoryFilter.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\actions\\index.js",[],"C:\\Users\\User\\Desktop\\bookstore\\src\\components\\Header.js",[]]
