+++ src/components/CategoryFilter.js
@@ -1,10 +1,19 @@
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import { CHANGE_FILTER } from '../actions';
+import styles from './CategoryFilter.module.css';
 
 const { v4: UuidV4 } = require('uuid');
 
 const CategoryFilter = props => {
-  const { handleFilterChange, filter } = props;
+  const { onChangeFilter, filter } = props;
+
+  const handleFilterChange = event => {
+    const {
+      target: { value },
+    } = event;
+    onChangeFilter(value);
+  };
 
   const categories = [
     'Action',
@@ -17,28 +26,22 @@ const CategoryFilter = props => {
   ];
 
   return (
-    <tr>
-      <td>
-        <label htmlFor="filter-categories">
-          Choose a category
-          <select
-            id="filter-categories"
-            name="category"
-            onChange={handleFilterChange}
-            value={filter}
-          >
-            <option value="" key={UuidV4()}>
-              All
-            </option>
-            {categories.map(category => (
-              <option value={category} key={UuidV4()}>
-                {category}
-              </option>
-            ))}
-          </select>
-        </label>
-      </td>
-    </tr>
+    <select
+      id="filter-categories"
+      className={styles.category_select}
+      name="category"
+      onChange={handleFilterChange}
+      value={filter}
+    >
+      <option value="" key={UuidV4()}>
+        All
+      </option>
+      {categories.map(category => (
+        <option value={category} key={UuidV4()}>
+          {category}
+        </option>
+      ))}
+    </select>
   );
 };
 
@@ -46,9 +49,13 @@ const mapStateToProps = state => ({
   filter: state.filt.filter,
 });
 
+const mapDispatchToProps = dispatch => ({
+  onChangeFilter: bookCategory => dispatch(CHANGE_FILTER(bookCategory)),
+});
+
 CategoryFilter.propTypes = {
-  handleFilterChange: PropTypes.func.isRequired,
+  onChangeFilter: PropTypes.func.isRequired,
   filter: PropTypes.string.isRequired,
 };
 
-export default connect(mapStateToProps)(CategoryFilter);
+export default connect(mapStateToProps, mapDispatchToProps)(CategoryFilter);
