+++ src/components/CategoryFilter.module.css
@@ -1,0 +1,5 @@
+.category_select {
+  width: 90px;
+  height: 20px;
+  margin-top: 15px;
+}
