+++ src/components/Header.module.css
@@ -14,10 +14,10 @@
 .bookstore {
   width: 240px;
   height: 37px;
-  margin: 6px 48px 2px 0;
+  margin: 12px 48px 2px 0;
   font-family: Montserrat;
-  font-size: 30px;
-  font-weight: bold;
+  font-size: 25px;
+  font-weight: bolder;
   font-stretch: normal;
   font-style: normal;
   line-height: normal;
