+++ src/containers/BookList.js
@@ -1,26 +1,15 @@
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import Book from '../components/Book';
-import { REMOVE_BOOK, CHANGE_FILTER } from '../actions';
-import CategoryFilter from '../components/CategoryFilter';
+import { REMOVE_BOOK } from '../actions';
 
 const { v4: UuidV4 } = require('uuid');
 
 const BookList = props => {
-  const {
-    books, filter, onDeleteBook, onChangeFilter,
-  } = props;
+  const { books, filter, onDeleteBook } = props;
   const handleClick = bookId => {
     onDeleteBook(bookId);
   };
-
-  const handleFilterChange = event => {
-    const {
-      target: { value },
-    } = event;
-    onChangeFilter(value);
-  };
-
   return (
     <table>
       <tbody>
@@ -40,9 +29,6 @@ const BookList = props => {
               handleClick={() => handleClick(book.id)}
             />
           ))}
-        <CategoryFilter
-          handleFilterChange={event => handleFilterChange(event)}
-        />
       </tbody>
     </table>
   );
@@ -55,14 +41,12 @@ const mapStateToProps = state => ({
 
 const mapDispatchToProps = dispatch => ({
   onDeleteBook: bookId => dispatch(REMOVE_BOOK(bookId)),
-  onChangeFilter: bookCategory => dispatch(CHANGE_FILTER(bookCategory)),
 });
 
 BookList.propTypes = {
   books: PropTypes.arrayOf(PropTypes.object).isRequired,
   filter: PropTypes.string.isRequired,
   onDeleteBook: PropTypes.func.isRequired,
-  onChangeFilter: PropTypes.func.isRequired,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(BookList);
